requirements.txt
camelot-py==0.11.0
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
et-xmlfile==1.1.0
ghostscript==0.7
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
pandas==2.2.0
pdfminer.six==20231228
pycparser==2.21
pypdf==4.0.1
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
tabulate==0.9.0
tzdata==2023.4
